@@ -39,14 +39,15 @@
         "@types/react-bootstrap": "^0.32.21",
         "@types/react-dom": "^16.9.8",
         "@types/react-redux": "^7.1.9",
-        "@types/react-router": "^5.1.7",
-        "@types/react-router-dom": "^5.1.5",
+        "@types/react-router": "^5.1.18",
+        "@types/react-router-dom": "^5.3.3",
         "@types/recompose": "^0.30.7",
         "@types/simple-peer": "^9.6.1",
         "@types/socket.io": "^2.1.8",
         "@types/socket.io-client": "^1.4.33",
         "@types/underscore": "^1.10.0",
         "@types/uuid": "^8.0.0",
+        "@types/webspeechapi": "0.0.29",
         "add": "^2.0.6",
         "apollo-upload-client": "^14.1.3",
         "awesome-typescript-loader": "^5.2.1",
@@ -87,7 +88,8 @@
         "react-google-login": "^5.1.20",
         "react-hooks-worker": "^0.9.0",
         "react-hotkeys": "^2.0.0",
-        "react-router-dom": "^5.2.0",
+        "react-router": "^6.2.2",
+        "react-router-dom": "^6.2.2",
         "recompose": "^0.30.0",
         "remixicon": "^2.5.0",
         "reset-css": "^4.0.1",
@@ -190,7 +192,7 @@
         "ts-node": "^8.10.2",
         "tslint-config-prettier": "^1.18.0",
         "tslint-plugin-prettier": "^2.3.0",
-        "typescript": "^3.9.5",
+        "typescript": "^4.6.3",
         "waait": "^1.0.5",
         "web-audio-test-api": "^0.5.2",
         "webpack": "^4.43.0",
